update chart all

+++ src/main/java/kr/co/iei/news/controller/NewsController.java
@@ -1,5 +1,6 @@
 package kr.co.iei.news.controller;
 
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 
@@ -9,10 +10,12 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.SessionAttribute;
 import org.springframework.web.multipart.MultipartFile;
+import kr.co.iei.customer.controller.CustomerController;
 import kr.co.iei.member.model.vo.Member;
 import kr.co.iei.news.model.service.NewsService;
 import kr.co.iei.news.model.vo.Discount;
@@ -26,6 +29,8 @@ import kr.co.iei.util.FileUtil;
 @RequestMapping("news")
 public class NewsController {
 
+    private final CustomerController customerController;
+
 
 	
 	@Autowired
@@ -36,6 +41,10 @@ public class NewsController {
 	
 	@Autowired
 	private FileUtil fileUtil;
+
+    NewsController(CustomerController customerController) {
+        this.customerController = customerController;
+    }
 	
 	@GetMapping(value="/list")
 	public String newsList(int noticeReqPage, String tab, Model model) {
@@ -48,15 +57,16 @@ public class NewsController {
 		model.addAttribute("totalCount", totalCount);
 		model.addAttribute("tab", tab);
 		
+		//배너 정보
 		Poster banner = newsService.selectNewsPoster();
 		model.addAttribute("banner", banner);
 		
 		List<News> newsAllList = newsService.selectAllNewsBanner();
-		System.out.println(newsAllList);
 		model.addAttribute("newsAllList", newsAllList);
 		
 		News newsModal = newsService.selectOneNewsModal();
 		model.addAttribute("newsModal", newsModal);
+		model.addAttribute("modal", 1);
 		return "news/list";
 	}
 	
@@ -90,12 +100,13 @@ public class NewsController {
 		System.out.println(discountPrice);
 		int result = newsService.insertNews(news, productList, discountType, discountPrice);		
 		
-		return "redirect:/news/list?noticeReqPage=1&tab=all";
+		return "redirect:/news/list?noticeReqPage=1&tab=all&modal=1";
 	}
 	
 	
 	@GetMapping(value="updateNews")
 	public String updateNews(int newsNo, Model model) {
+		
 		News news = newsService.selectOneNews(newsNo);
 		model.addAttribute("news", news);
 		
@@ -103,15 +114,23 @@ public class NewsController {
 		model.addAttribute("product", product);
 		
 		List discount = newsService.selectAllDiscount(newsNo);
-		Discount d = (Discount)discount.get(0);
-		
-		if(d.getDiscountPercent()== 0) {
-			model.addAttribute("discountPrice",d.getDiscountPrice());
+		System.out.println(discount);
+		if(!discount.isEmpty()) {
+			Discount d = (Discount)discount.get(0);
+			if(d.getDiscountPercent()== 0) {
+				d.setDiscountType("Price");
+				model.addAttribute("discountPrice",d.getDiscountPrice());	
+				
+			} else {
+				d.setDiscountType("Percent");
+				model.addAttribute("discountPrice",d.getDiscountPercent());
+			}
 		} else {
-			model.addAttribute("discountPrice",d.getDiscountPercent());
+			model.addAttribute("discountPrice",0);
+			model.addAttribute("discountType",0);
 		}
-		model.addAttribute("discount", discount);
 		
+		model.addAttribute("discount", discount);
 		return "news/updateNews";
 	}
 	
@@ -140,9 +159,13 @@ public class NewsController {
 	@GetMapping(value="/noticeWrite")
 	public String noticeWrite(Notice notice, @SessionAttribute(required = false) Member member) {
 		int memberNo = member == null ? 0 : member.getMemberNo();
+		if(memberNo != 0 )  {
+			notice.setMemberNo(memberNo);
+			int result = newsService.insertNotice(notice);			
+			
+		}
 		
-		int result = newsService.insertNotice(notice);
-		return "redirect:/news/list?noticeReqPage=1&tab=all";
+		return "redirect:/news/list?noticeReqPage=1&tab=all&modal=1";
 	}
 	
 	@GetMapping(value="/noticeView")
@@ -181,9 +204,13 @@ public class NewsController {
 	
 	@GetMapping(value="deleteNews")
 	public String deleteNews(int newsNo) {
-		int result = newsService.deleteNews(newsNo);
+		Poster posterMain = newsService.selectMainPoster();
+		Poster posterNews = newsService.selectNewsPoster();
+		if(newsNo !=posterMain.getNewsNo() && newsNo != posterNews.getNewsNo()) {
+			int result = newsService.deleteNews(newsNo);
+		} 
 		
-		return "redirect:/news/list?noticeReqPage=1&tab=all";
+		return "redirect:/news/list?noticeReqPage=1&tab=all&modal=1";
 	}
 	
 	@GetMapping(value="bannerWrite")
@@ -225,5 +252,54 @@ public class NewsController {
 		int result = newsService.changeModal(newsNo);		
 	}
 	
+	@PostMapping(value="bannerChange")
+	public String bannerChange(int posterNo, int posterNoNews) {
+		int result = newsService.changeBannerMain(posterNo);
+		result += newsService.changeBannerNews(posterNoNews);
+		return "redirect:/";
+	}
+	
+	@GetMapping(value="bannerDelete")
+	public String bannerDelete(Model model) {
+		List posterList = newsService.selectAllPoster();
+		model.addAttribute("posterList",posterList);
+		
+		return "news/bannerDelete";
+	}
+	@GetMapping(value="deleteBannerComplete")
+	public String deleteBannerComplete(String posterNo) {
+		
+		
+		String[] posterNoList;
+		if(posterNo != null) {	
+			posterNoList = posterNo.split(",");
+		} else {
+			posterNoList = null;
+		}
+		int result = newsService.deleteBanner(posterNoList);		
+		return  "redirect:/news/list?noticeReqPage=1&tab=all&modal=1";
+	}
+	@ResponseBody
+	@GetMapping(value="noticeDelete")
+	public void  noticeDelete(int noticeNo) {
+		int result = newsService.deleteNotice(noticeNo);
+	}
+	
+	@GetMapping(value="noticeUpdateFrm")
+	public String noticeUpdateFrm(int noticeNo, Model model) {
+		Notice notice = newsService.selectOneNotice(noticeNo);
+		model.addAttribute("notice", notice);
+		
+		return "news/noticeUpdateFrm";
+	}
+	
+	@GetMapping(value="noticeUpdate")
+	public String noticeupdate(Notice notice, @SessionAttribute(required = false) Member member) {
+		if(member != null) {
+			notice.setMemberNo(member.getMemberNo());
+			int result = newsService.updateNotice(notice);
+		}
+		return  "redirect:/news/list?noticeReqPage=1&tab=all&modal=1";
+	}
 }
